--- feelings/src/App.tsx
@@ -1,19 +1,0 @@
-// import styled from "styled-components";
-
-function App() {
-    return (
-        <>
-            <p>how are you feeling today?</p>
-            <form>
-                <input
-                    type="text"
-                    placeholder="type here + press enter"
-                    id="inputbox"
-                    name="feeling"
-                />
-            </form>
-        </>
-    );
-}
-
-export default App;
