+++ feelings/web/package.json
@@ -13,6 +13,7 @@
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "@types/styled-components": "^5.1.11",
+    "axios": "^0.21.1",
     "express": "^4.17.1",
     "mikro-orm": "^4.5.7",
     "react": "^17.0.2",
